# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "LedgerLine backend API.",
        "title": "LedgerLine API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Tim LedgerLine",
            "email": "dev@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/forgot-password": {
            "post": {
                "description": "Sends an OTP code if the address is registered; the answer is the same either way",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Request a password reset OTP",
                "parameters": [
                    {
                        "description": "Destination email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.ForgotPasswordRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Exchanges email and password for an access + refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.LoginRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.TokenResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Profile of the currently logged-in user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.ProfileResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Issues a new access token using the refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RefreshTokenRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.TokenResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Creates a new account with the default \"user\" role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RegisterRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.UserResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "description": "Changes the password using the reset token from the OTP verification step",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Change password with a reset token",
                "parameters": [
                    {
                        "description": "Reset token + new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.ResetPasswordRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            }
        },
        "/auth/verify-otp": {
            "post": {
                "description": "Checks the OTP code before the user may change their password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify the password reset OTP",
                "parameters": [
                    {
                        "description": "Email + OTP code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.VerifyOTPResetPasswordRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.ResetTokenResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.ForgotPasswordRequestDTO": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "budi@example.com"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.LoginRequestDTO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "budi@example.com"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8,
                    "example": "Rahasia123!"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.MenuAccessResponseDTO": {
            "type": "object",
            "properties": {
                "can_approve": {
                    "type": "boolean",
                    "example": false
                },
                "can_create": {
                    "type": "boolean",
                    "example": true
                },
                "can_delete": {
                    "type": "boolean",
                    "example": true
                },
                "can_read": {
                    "type": "boolean",
                    "example": true
                },
                "can_update": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.MenuResponseDTO": {
            "type": "object",
            "properties": {
                "access": {
                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.MenuAccessResponseDTO"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.MenuResponseDTO"
                    }
                },
                "code": {
                    "type": "string",
                    "example": "transactions"
                },
                "icon": {
                    "type": "string",
                    "example": "swap"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Transactions"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.ProfileResponseDTO": {
            "type": "object",
            "properties": {
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.MenuResponseDTO"
                    }
                },
                "user": {
                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.UserResponseDTO"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RefreshTokenRequestDTO": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RegisterRequestDTO": {
            "type": "object",
            "required": [
                "email",
                "full_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "budi@example.com"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Budi Santoso"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8,
                    "example": "Rahasia123!"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.ResetPasswordRequestDTO": {
            "type": "object",
            "required": [
                "confirm_new_password",
                "new_password",
                "reset_token"
            ],
            "properties": {
                "confirm_new_password": {
                    "type": "string",
                    "example": "RahasiaBaru123!"
                },
                "new_password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8,
                    "example": "RahasiaBaru123!"
                },
                "reset_token": {
                    "type": "string",
                    "example": "abc123"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.ResetTokenResponseDTO": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer",
                    "example": 600
                },
                "reset_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                }
            }
        },
...
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.TokenResponseDTO": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                },
                "expires_in": {
                    "type": "integer",
                    "example": 900
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
...
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.UserResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-01-02T15:04:05Z"
                },
                "email": {
                    "type": "string",
                    "example": "budi@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "id": {
                    "type": "string",
                    "example": "6f1e2b7e-2c8a-4c1f-9f3e-6a0f1c2d3e4b"
                },
                "role": {
                    "type": "string",
                    "example": "User"
                },
                "role_id": {
                    "type": "string",
                    "example": "00000000-0000-0000-0000-000000000002"
                },
                "status": {
                    "type": "string",
                    "example": "enabled"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-01-02T15:04:05Z"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.VerifyOTPResetPasswordRequestDTO": {
            "type": "object",
            "required": [
                "email",
                "otp"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "budi@example.com"
                },
                "otp": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_pkg_response.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "USER_NOT_FOUND"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.FieldError"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "something went wrong"
                },
                "request_id": {
                    "type": "string",
                    "example": "9b2c1f2e-6f0a-4a1e-9c7d-2f8b0a1c3d4e"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_pkg_response.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "type": "string",
                    "example": "email must be a valid email address"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "per_page": {
                    "type": "integer",
                    "example": 10
                },
                "total_items": {
                    "type": "integer",
                    "example": 42
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_pkg_response.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Fill with: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}